refactor: extract overview tab to views/OverviewView.jsx + child tests
Strangler-fig extraction #4 of the sprint (#74, part of #52) — the
second-largest tab (~1,370 lines). Behaviour unchanged; code moved, not
rewritten.

- new views/OverviewView.jsx takes 11 props (latest, loadData,
  loggedSessions, latestErg, latestSquat, totalErgDist, totalSessions,
  ftp, isWide, tsbColor, nowTick); no local state
- monolith renders {view === 'overview' && <OverviewView .../>} and sheds
  ~24 now-orphaned imports (all of recharts, LoadTooltip, the analysis
  utils, several constants)
- folds in #53: real tests for the children the overview pulls in —
  WorkoutItem, LoadTooltip, LogEntry — each gated at the commercial
  80/80/70; OverviewView itself gated too (100/100/70.5)
- global coverage floor ratcheted 48/46/40 -> 52/50/48 (measured ~63/61/64)

Monolith: 5,834 -> 4,456 lines.
